yukawa: Unify fstab files to fix issues w/ adb remount
The overlayfs remount handling of superpartitions expects the
fstab to have all the partitions in it.

This means we can't use the split between the ramdisk fstab
and the vendor fstab that we have been previously.

So this patch merges the two fstabs together.

With this change "adb root; adb remount" works now.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Iad408d5d3b1e4c9291a947cd44ee65f4aaef5c45