DO NOT MERGE: yukawa: Use uncompressed kernel
When building android-mainline kernel, we cannot rely on lz4 because the
compression algorithm used by the kernel is not the same as
the one implemented in U-Boot.

The kernel uses the legacy (-l) option, which is incompatible with with U-Boot.
To be compatible with U-Boot we need the Image.lz4 to be
generated with:

$ lz4 -f -12 --favor-decSpeed /arch/arm64/boot/Image /arch/arm64/boot/Image.lz4

Disable compression for now to avoid errors.

Note: this breaks all previous kernel builds, which is why this is marked as DO
NOT MERGE.

Change-Id: I3cfb1005b4ee9be638f7afa2cc08cc0dd3ecd524
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>